The script is more or less functional
The script file is a complete mess, but it's more or less functional, still need to comment everything, move things around to make it more legible and to create the function that iterates trough all the movies in the list until it gets to 10k reviews.
BUT it does function, it's slow as hell but it's functional.